Connect postgresql - calculate sim

+++ foodbike/analytics/migrations/0001_initial.py
@@ -25,7 +25,7 @@ class Migration(migrations.Migration):
                 ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('user_id', models.CharField(max_length=128)),
                 ('food_id', models.IntegerField()),
-                ('rating', models.IntegerField()),
+                ('rating', models.DecimalField(decimal_places=2, max_digits=4)),
                 ('rating_timestamp', models.DateTimeField()),
                 ('type', models.CharField(default=b'explicit', max_length=8)),
             ],
